+++ samples/lib/widgets/popover_menu.dart
@@ -31,9 +31,10 @@ class _State extends StatefulSampleState<PopoverMenuPage> with SingleTickerProvi
         title: const Text('Edit Notes'),
         actions: [
           FPopoverMenu(
-            controller: controller,
+            popoverController: controller,
             menuAnchor: Alignment.topRight,
             childAnchor: Alignment.bottomRight,
+            maxHeight: 200,
             menu: [
               FTileGroup(
                 children: [
@@ -45,6 +46,11 @@ class _State extends StatefulSampleState<PopoverMenuPage> with SingleTickerProvi
                   FTile(
                     prefixIcon: FIcon(FAssets.icons.paperclip),
                     title: const Text('Add attachments'),
+                    onPress: () {},
+                  ),
+                  FTile(
+                    prefixIcon: FIcon(FAssets.icons.qrCode),
+                    title: const Text('Scan Document'),
                     onPress: () {},
                   ),
                 ],